some refactoring work
