+++ models/pstn.py
@@ -2,6 +2,8 @@ from __future__ import print_function
 
 import torch
 import torch.nn as nn
+from torch import distributions
+from utils.transformers import init_transformer
 
 
 class PSTN(nn.Module):
@@ -12,6 +14,12 @@ class PSTN(nn.Module):
         self.num_classes = opt.num_classes
         self.num_param = opt.num_param
         self.N = opt.N
+        self.transformer, self.theta_dim = init_transformer(opt)
+
+        self.train_samples = opt.train_samples
+        self.test_samples = opt.test_samples
+        self.alpha_p = opt.alpha_p
+        self.beta_p = opt.beta_p
 
         # Spatial transformer localization-network
         self.init_localizer(opt)
@@ -21,28 +29,10 @@ class PSTN(nn.Module):
         self.init_model_weights(opt)
 
     def init_localizer(self, opt):
-        if opt.dataset.lower() in ['cub']:
-            from .cublocalizer import CubPSTN as PSTN
-        elif opt.dataset.lower() in ['celeba']:
-            from .celebalocalizer import CelebaPSTN as PSTN
-        elif opt.dataset.lower() in ['mnist', 'random_placement_mnist']:
-            from .mnistlocalizer import MnistPSTN as PSTN
-        elif opt.dataset in opt.TIMESERIESDATASETS:
-            from .timeserieslocalizer import TimeseriesPSTN as PSTN
-
-        self.pstn = PSTN(opt)
+        raise NotImplementedError
 
     def init_classifier(self, opt):
-        if opt.dataset.lower() in ['cub']:
-            from .cubclassifier import CubClassifier as Classifier
-        elif opt.dataset.lower() in ['celeba']:
-            from .celebaclassifier import CelebaClassifier as Classifier
-        elif opt.dataset.lower() in ['mnist', 'random_placement_mnist']:
-            from .mnistclassifier import MnistClassifier as Classifier
-        elif opt.dataset in opt.TIMESERIESDATASETS:
-            from .timeseriesclassifier import TimeseriesClassifier as Classifier
-
-        self.classifier = Classifier(opt)
+        raise NotImplementedError
 
     def init_model_weights(self, opt):
         self.pstn.fc_loc_mu[-1].weight.data.zero_()
@@ -54,24 +44,24 @@ class PSTN(nn.Module):
             if self.num_param == 2:
                 # We initialize bounding boxes with tiling
                 bias = torch.tensor([[-1, -1], [1, -1], [1, 1], [-1, 1]], dtype=torch.float) * 0.5
-                self.pstn.fc_loc_mu[-1].bias.data.copy_(bias[:self.N].view(-1))
+                self.fc_loc_mu[-1].bias.data.copy_(bias[:self.N].view(-1))
             elif self.num_param == 3:
-                self.pstn.fc_loc_mu[-1].bias.data.copy_(torch.tensor([1, 0, 0] * self.N, dtype=torch.float))
+                self.fc_loc_mu[-1].bias.data.copy_(torch.tensor([1, 0, 0] * self.N, dtype=torch.float))
             elif self.num_param == 4:
-                self.pstn.fc_loc_mu[-1].bias.data.copy_(torch.tensor([1, 1, 0, 0] * self.N, dtype=torch.float))
+                self.fc_loc_mu[-1].bias.data.copy_(torch.tensor([1, 1, 0, 0] * self.N, dtype=torch.float))
             elif self.num_param == 5:
-                self.pstn.fc_loc_mu[-1].bias.data.copy_(torch.tensor([0, 1, 1, 0, 0] * self.N, dtype=torch.float))
+                self.fc_loc_mu[-1].bias.data.copy_(torch.tensor([0, 1, 1, 0, 0] * self.N, dtype=torch.float))
             elif self.num_param == 6:
-                self.pstn.fc_loc_mu[-1].bias.data.copy_(torch.tensor([1, 0, 0,
+                self.fc_loc_mu[-1].bias.data.copy_(torch.tensor([1, 0, 0,
                                                                       0, 1, 0] * self.N, dtype=torch.float))
             # initialize beta network
-            self.pstn.fc_loc_beta[-2].weight.data.zero_() # TODO: check that this is still a good init
-            self.pstn.fc_loc_beta[-2].bias.data.copy_(
+            self.fc_loc_beta[-2].weight.data.zero_() # TODO: check that this is still a good init
+            self.fc_loc_beta[-2].bias.data.copy_(
                 torch.tensor([-5], dtype=torch.float).repeat(self.num_param * self.N))
 
         elif opt.transformer_type == 'diffeomorphic':
             # initialize param's as identity, default ok for variance in this case
-            self.pstn.fc_loc_mu[-1].bias.data.copy_(
+            self.fc_loc_mu[-1].bias.data.copy_(
                 torch.tensor([1e-5], dtype=torch.float).repeat(self.pstn.theta_dim))
             self.pstn.fc_loc_std[-2].weight.data.zero_()
 
@@ -84,11 +74,12 @@ class PSTN(nn.Module):
         batch_size = x.shape[0]
 
         # get output for pstn module
-        x, thetas, beta = self.pstn(x) # fix alpha for now 
+        x, thetas, beta = self.forward_localizer(x) # fix alpha for now 
         # x, thetas, (theta_mu, theta_sigma) = self.pstn(x)
 
         # make classification based on pstn output
-        x = self.classifier(x)
+        x = self.forward_classifier(x)
+
         # format according to number of samples
         x = torch.stack(x.split([batch_size] * self.pstn.S))
         x = x.view(self.pstn.S, batch_size * self.num_classes)
@@ -103,3 +94,38 @@ class PSTN(nn.Module):
             x = x.view(batch_size, self.num_classes)
 
         return x, thetas, beta
+
+    def forward_localizer(self, x):
+        self.S = self.train_samples if self.training else self.test_samples
+        batch_size, c, w, h = x.shape
+        # shared localizer
+        xs = self.localization(x)
+        xs = xs.view(batch_size, -1)
+        # estimate mean and variance regressor
+        theta_mu = self.fc_loc_mu(xs)
+        beta = self.fc_loc_beta(xs)
+
+        # repeat x in the batch dim so we avoid for loop
+        # (this doesn't do anything for N=1)
+        x = x.unsqueeze(1).repeat(1, self.N, 1, 1, 1).view(self.N * batch_size, c, w, h)
+        theta_mu_upsample = theta_mu.view(batch_size * self.N, self.theta_dim) # mean is the same for all S: [bs * N, theta_dim]
+        beta_upsample = beta.view(batch_size * self.N, self.theta_dim) # variance is also the same, difference comes in through sampling
+        alpha_upsample = self.alpha_p * torch.ones_like(theta_mu_upsample) # upsample scalar alpha
+
+        # make the T-dist object and sample it here? 
+        # it's apparently ok to generate distribution anew in each forward pass (e.g. https://github.com/kampta/pytorch-distributions/blob/master/gaussian_vae.py
+        # maybe we could do this more efficiently because of the independence assumptions within theta? 
+        T_dist = distributions.studentT.StudentT(df= 2* alpha_upsample, loc=theta_mu_upsample, scale=torch.sqrt(beta_upsample / alpha_upsample))
+        theta_samples = T_dist.rsample([self.S]) # shape: [self.S, batch_size, self.theta_dim]
+        theta_samples = theta_samples.view([self.S * batch_size, self.theta_dim])
+
+        # repeat for the number of samples
+        x = x.repeat(self.S, 1, 1, 1)
+        x = x.view([self.S * batch_size, c, w, h])
+        x = self.transformer(x, theta_samples)
+
+        # theta samples: [S, bs, nr_params]
+        return x, theta_samples, (theta_mu, beta)
+
+    def forward_classifier(self, x):
+        return self.classifier(x)
